✨ 環境設定を更新し、不要なコンポーネントを削除。新しいSummaryPropsインターフェースを追加し、CustomButtonコンポーネントを作成。データ表示機能を強化し、型定義を改善。

+++ backend/app/schema/ix_non_fraction.py
@@ -36,14 +36,14 @@ class IxNonFractionPublic(SQLModel):
     update_date: datetime
     head_item_key: Optional[str]
     context: str
-    decimals: Optional[Decimal]
+    decimals: Optional[float]
     format: Optional[str]
     name: str
-    scale: Optional[Decimal]
+    scale: Optional[float]
     sign: Optional[str]
     unit_ref: Optional[str]
     xsi_nil: Optional[bool]
-    numeric: Optional[Decimal]
+    numeric: Optional[float]
     report_type: Optional[str]
     ixbrl_role: Optional[str]
     source_file_id: Optional[str]
